+++ account_bank_statement_import_camt/test_files/test-camt053.xml
