+++ account_bank_statement_import_mt940_nl_ing/mt940.py
@@ -19,7 +19,7 @@
 #
 ##############################################################################
 import re
-from openerp.addons.bank_statement_parse_mt940.mt940 import (
+from openerp.addons.account_bank_statement_import_mt940_base.mt940 import (
     MT940, str2amount, get_subfields, handle_common_subfields)
