+++ account_bank_statement_import_mt940_nl_rabo/tests/test_import_bank_statement.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-"""Run test to import MT940 IBAN ING import."""
+"""Run test to import MT940 IBAN RABO import."""
 ##############################################################################
 #
 #    Copyright (C) 2015 Therp BV <http://therp.nl>.
-#
 #    All other contributions are (C) by their respective contributors
 #
 #    This program is free software: you can redistribute it and/or modify
@@ -25,29 +24,23 @@ from openerp.addons.account_bank_statement_import.tests import (
 
 
 class TestImport(TestStatementFile):
-    """Run test to import MT940 ING import."""
-
-    def test_old_statement_import(self):
-        """Test correct creation of single statement from old format."""
-        self._test_statement_import(
-            'bank_statement_parse_nl_ing_mt940', 'test-ing-old.940',
-            'NL77INGB0574908765-2014-01-20',
-            start_balance=662.23, end_balance=564.35
-        )
+    """Run test to import MT940 RABO import."""
 
     def test_statement_import(self):
         """Test correct creation of single statement."""
         transactions = [
             {
-                'remote_account': 'NL32INGB0000012345',
-                'transferred_amount': 1.56,
-                'value_date': '2014-02-20',
-                'ref': 'EV12341REP1231456T1234',
+                'remote_account': 'NL66RABO0160878799',
+                'transferred_amount': 400.00,
+                'value_date': '2014-01-02',
+                'ref': 'NONREF',
             },
         ]
+        # statement name is account number + '-' + date of last 62F line:
         self._test_statement_import(
-            'bank_statement_parse_nl_ing_mt940', 'test-ing.940',
-            'NL77INGB0574908765-2014-02-20',
-            start_balance=662.23, end_balance=564.35,
+            'account_bank_statement_import_mt940_nl_rabo', 'test-rabo.swi',
+            'NL34RABO0142623393-2014-01-07',
+            local_account='NL34RABO0142623393',
+            start_balance=4433.52, end_balance=4798.91,
             transactions=transactions
         )
