--- bank_statement_parse/README.rst
@@ -1,49 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
-Bank Statement Parse
-====================
-
-This module makes it easy to write parsers for bank statement files, by
-providing common functionality. The module is especially usefull for
-converting the parsers written for previous versions of OpenERp / Odoo.
-
-Known issues / Roadmap
-======================
-
-* None
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/bank-statement-import/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/OCA/bank-statement-import/issues/new?body=module:%20account_bank_statement_import%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-
-Credits
-=======
-
-Contributors
-------------
-
-* Stefan Rijnhart <srijnhart@therp.nl>
-* Ronald Portier <rportier@therp.nl>
-
-Maintainer
-----------
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-This module is maintained by the OCA.
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-To contribute to this module, please visit http://odoo-community.org.
-This module should make it easy to migrate bank statement import
-modules written for earlies versions of Odoo/OpenERP.
